// src/types/tec.ts
export interface TECArticle {
  sh10Code: string;
  designation: string;
  us: string;
  dd: number;
  rsta: number;
  pcs: number;
  pua: number;
  pcc: number;
  rrr: number;
  rcp: number;
  cumulSansTVA: number;
  cumulAvecTVA: number;
  tva: number;
  sh6Code: string;
  tub: string;
  dus: string;
  dud: string;
  tcb: string;
  tsm: string;
  tsb: string;
  psv: string;
  tai: string;
  tab: string;
  tuf: string;
} 